슬라이드,사이드바 분리
