+++ src/main/moivo_react/src/assets/css/product_board.module.css
@@ -1,3 +1,5 @@
+/* product_board.module.css */
+
 /* 전체 컨테이너 스타일 */
 .container {
   width: 100%;
@@ -24,7 +26,7 @@
 .bar {
   height: 5px;
   width: 35px;
-  background-color: #EF8354;
+  background-color: #353433;
   margin: 6px auto;
   transition: all 0.5s ease;
 }
@@ -48,7 +50,7 @@
   position: fixed;
   top: 0;
   right: 0;
-  background-color: #EF8354;
+  background-color: rgb(180, 165, 155);
   transform: translateX(100%);
   transition: transform 0.7s ease-in-out;
   z-index: 1000;
@@ -74,15 +76,6 @@
   text-decoration: none;
   font-size: 1.5em;
   display: block;
-}
-
-/* 로그인 아이콘 */
-.ua {
-  position: fixed;
-  bottom: 20px;
-  left: 50%;
-  transform: translateX(-50%);
-  z-index: 1100;
 }
 
 /* 슬라이더 */
@@ -122,7 +115,7 @@
 
 /* 푸터 */
 .mainFooter {
-  background: #2F2E2C;
+  background: rgb(170, 166, 162);
   padding: 2rem 0;
 }
